회원 좋아요 목록 찾기 기능 구현, 테스트 코드 작성

+++ src/main/java/com/webtoon/member/dto/request/MemberCharge.java
@@ -11,10 +11,10 @@ import javax.validation.constraints.Min;
 public class MemberCharge {
 
     @Min(value = 0, message = "0코인 이상을 충전해주세요")
-    private Integer chargeAmount;
+    private Long chargeAmount;
 
     @Builder
-    public MemberCharge(Integer chargeAmount) {
+    public MemberCharge(Long chargeAmount) {
         this.chargeAmount = chargeAmount;
     }
 }
